+++ pulsar-plugins/pulsar-protocol/src/main/kotlin/ai/platon/pulsar/protocol/browser/driver/BrowserInstance.kt
@@ -1,9 +1,15 @@
 package ai.platon.pulsar.protocol.browser.driver
 
-import ai.platon.pulsar.browser.driver.chrome.*
-import ai.platon.pulsar.common.config.CapabilityTypes
+import ai.platon.pulsar.browser.driver.chrome.ChromeDevtoolsOptions
+import ai.platon.pulsar.browser.driver.chrome.ChromeLauncherV2
+import ai.platon.pulsar.browser.driver.chrome.LauncherConfig
 import ai.platon.pulsar.crawl.fetch.privacy.BrowserInstanceId
 import ai.platon.pulsar.protocol.browser.driver.chrome.ChromeDevtoolsDriver
+import com.github.kklisura.cdt.launch.ChromeArguments
+import com.github.kklisura.cdt.launch.ChromeLauncher
+import com.github.kklisura.cdt.services.ChromeDevToolsService
+import com.github.kklisura.cdt.services.ChromeService
+import com.github.kklisura.cdt.services.types.ChromeTab
 import org.slf4j.LoggerFactory
 import java.util.concurrent.ConcurrentLinkedQueue
 import java.util.concurrent.atomic.AtomicBoolean
@@ -14,8 +20,8 @@ import java.util.concurrent.atomic.AtomicInteger
  * Copyright @ 2013-2017 Platon AI. All rights reserved
  */
 class BrowserInstance(
-        val launcherConfig: LauncherConfig,
-        val launchOptions: ChromeDevtoolsOptions
+    val launcherConfig: LauncherConfig,
+    val launchOptions: ChromeDevtoolsOptions
 ): AutoCloseable {
 
     /**
@@ -27,9 +33,9 @@ class BrowserInstance(
     val proxyServer get() = launchOptions.proxyServer
 
     val numTabs = AtomicInteger()
-    lateinit var launcher: ChromeLauncher
-    lateinit var chrome: RemoteChrome
-    val devToolsList = ConcurrentLinkedQueue<RemoteDevTools>()
+    lateinit var launcher: ChromeLauncherV2
+    lateinit var chrome: ChromeService
+    val devToolsList = ConcurrentLinkedQueue<ChromeDevToolsService>()
 
     private val log = LoggerFactory.getLogger(BrowserInstance::class.java)
     private val launched = AtomicBoolean()
@@ -41,7 +47,20 @@ class BrowserInstance(
         synchronized(ChromeLauncher::class.java) {
             if (launched.compareAndSet(false, true)) {
                 val shutdownHookRegistry = ChromeDevtoolsDriver.ShutdownHookRegistry()
-                launcher = ChromeLauncher(config = launcherConfig, shutdownHookRegistry = shutdownHookRegistry)
+//                launcher = ChromeLauncherKK(
+//                    processLauncher = ProcessLauncherImpl(),
+//                    environment = ChromeLauncherKK.Environment { System.getenv(it) },
+//                    shutdownHookRegistry = ChromeLauncherKK.RuntimeShutdownHookRegistry(),
+//                    configuration = ChromeLauncherConfiguration())
+
+                val args = ChromeArguments.defaults(false) // Sets the correct arguments: enable-logging and logging level
+                    .enableLogging("stderr")
+                    .userDataDir("/home/vincent/.pulsar/browser/chrome/prototype/google-chrome")
+                    .additionalArguments("v", "1")
+                    .build()
+                launcher = ChromeLauncherV2(config = launcherConfig)
+
+//                chrome = launcher.launch(launchOptions)
                 chrome = launcher.launch(launchOptions)
             }
         }
@@ -59,8 +78,9 @@ class BrowserInstance(
 
     @Synchronized
     @Throws(Exception::class)
-    fun createDevTools(tab: ChromeTab, config: DevToolsConfig): RemoteDevTools {
-        val devTools= chrome.createDevTools(tab, config)
+    fun createDevTools(tab: ChromeTab): ChromeDevToolsService {
+        val devTools= chrome.createDevToolsService(tab)
+        // chrome.createDevTools(tab, config)
         devToolsList.add(devTools)
         return devTools
     }
@@ -81,7 +101,7 @@ class BrowserInstance(
                 it.waitUntilClosed()
             }
 
-            chrome.close()
+            // chrome.close()
             launcher.close()
 
             log.info("Launcher is closed | {}", id.display)